tried to fix the authentication bug, added login credentials for database into connection tring in .env file
