+++ .env
@@ -1,3 +1,3 @@
-MONGO_URI=mongodb+srv://bookwm2016:<db_password>@users.ib4gi.mongodb.net/?retryWrites=true&w=majority&appName=users
+MONGO_URI=mongodb+srv://bookwm2016:<Theworldis1!>@users.ib4gi.mongodb.net/?retryWrites=true&w=majority&appName=users
 DB_NAME=IQ-app
 DB_URL=mongodb+srv://<credentials>@users.ib4gi.mongodb.net/project-3
