+++ client/server/server.js
@@ -16,8 +16,6 @@ mongoose.connect(process.env.MONGO_URI, {
 });
 
 app.use('/auth', authRoutes);
-app.use('/users', userRoutes);
-
 
 app.listen(5000, () => {
   console.log('Server is running on port 5000');
