+++ server/routes/authRoutes.js
@@ -1,0 +1,47 @@
+const express = require('express');
+const bcrypt = require('bcryptjs');
+const jwt = require('jsonwebtoken');
+const User = require('../models/User');
+
+const router = express.Router();
+
+// Register route
+router.post('/register', async (req, res) => {
+  const { username, password } = req.body;
+  const existingUser = await User.findOne({ username });
+
+  if (existingUser) {
+    return res.status(400).json({ message: 'User already exists' });
+  }
+
+  const hashedPassword = await bcrypt.hash(password, 10);
+  const newUser = new User({ username, password: hashedPassword });
+  await newUser.save();
+  res.status(201).json({ message: 'User created' });
+});
+
+// Login route
+router.post('/login', async (req, res) => {
+  const { username, password } = req.body;
+  const user = await User.findOne({ username });
+
+  if (!user || !(await bcrypt.compare(password, user.password))) {
+    return res.status(400).json({ message: 'Incorrect login information!' });
+  }
+
+  const token = jwt.sign({ id: user._id }, 'secret');
+  res.status(200).json({ token, username: user.username });
+});
+
+// Update user info
+router.put('/user/:id', async (req, res) => {
+  const { username, password } = req.body;
+  const hashedPassword = await bcrypt.hash(password, 10);
+  await User.findByIdAndUpdate(req.params.id, {
+    username,
+    password: hashedPassword,
+  });
+  res.status(200).json({ message: 'User updated' });
+});
+
+module.exports = router;
